+++ commands/run.go
@@ -23,6 +23,7 @@ import (
 )
 
 type Run struct {
+	conf              string
 	label             string
 	in                string
 	out               string
@@ -50,22 +51,23 @@ const UDP_TIMEOUT = 5 * time.Second
 func (r *Run) flags() *flag.FlagSet {
 	flagset := flag.NewFlagSet("", flag.ExitOnError)
 
-	flagset.StringVar(&r.in, "in", "", "tunnel connection that accepts external requests e.g. udp/listen:0.0.0.0:60000 or tcp/client:101.102.103.104:54321")
-	flagset.StringVar(&r.out, "out", "", "tunnel connection that dispatches received requests e.g. udp/broadcast:255.255.255.255:60000 or tcp/server:0.0.0.0:54321")
-	flagset.StringVar(&r.lockfile, "lockfile", "", "(optional) name of lockfile used to prevent running multiple copies of the service. A default lockfile name is generated if none is supplied")
-	flagset.IntVar(&r.maxRetries, "max-retries", MAX_RETRIES, "Maximum number of times to retry failed connection. Defaults to -1 (retry forever)")
-	flagset.DurationVar(&r.maxRetryDelay, "max-retry-delay", MAX_RETRY_DELAY, "Maximum delay between retrying failed connections")
-	flagset.DurationVar(&r.udpTimeout, "udp-timeout", UDP_TIMEOUT, "Time limit to wait for UDP replies")
+	flagset.StringVar(&r.conf, "config", r.conf, "optional tunnel TOML configuration file")
+	flagset.StringVar(&r.in, "in", r.in, "tunnel connection that accepts external requests e.g. udp/listen:0.0.0.0:60000 or tcp/client:101.102.103.104:54321")
+	flagset.StringVar(&r.out, "out", r.out, "tunnel connection that dispatches received requests e.g. udp/broadcast:255.255.255.255:60000 or tcp/server:0.0.0.0:54321")
+	flagset.StringVar(&r.lockfile, "lockfile", r.lockfile, "(optional) name of lockfile used to prevent running multiple copies of the service. A default lockfile name is generated if none is supplied")
+	flagset.IntVar(&r.maxRetries, "max-retries", r.maxRetries, "Maximum number of times to retry failed connection. Defaults to -1 (retry forever)")
+	flagset.DurationVar(&r.maxRetryDelay, "max-retry-delay", r.maxRetryDelay, "Maximum delay between retrying failed connections")
+	flagset.DurationVar(&r.udpTimeout, "udp-timeout", r.udpTimeout, "Time limit to wait for UDP replies")
 
-	flagset.StringVar(&r.caCertificate, "ca-cert", "ca.cert", "File path for CA certificate PEM file (defaults to ca.cert)")
-	flagset.StringVar(&r.certificate, "cert", "", "File path for client/server TLS certificate PEM file (defaults to client.cert or server.cert)")
-	flagset.StringVar(&r.key, "key", "", "File path for client/server TLS key PEM file (defaults to client.key or server.key)")
-	flagset.BoolVar(&r.requireClientAuth, "client-auth", false, "Requires client authentication for TLS")
+	flagset.StringVar(&r.caCertificate, "ca-cert", r.caCertificate, "File path for CA certificate PEM file (defaults to ca.cert)")
+	flagset.StringVar(&r.certificate, "cert", r.certificate, "File path for client/server TLS certificate PEM file (defaults to client.cert or server.cert)")
+	flagset.StringVar(&r.key, "key", r.key, "File path for client/server TLS key PEM file (defaults to client.key or server.key)")
+	flagset.BoolVar(&r.requireClientAuth, "client-auth", r.requireClientAuth, "Requires client authentication for TLS")
 
-	flagset.StringVar(&r.html, "html", "./html", "HTML folder for HTTP/HTTPS connectors")
-	flagset.StringVar(&r.logLevel, "log-level", "info", "Sets the log level (debug, info, warn or error)")
-	flagset.BoolVar(&r.console, "console", false, "Runs as a console application rather than a service")
-	flagset.BoolVar(&r.debug, "debug", false, "Enables detailed debugging logs")
+	flagset.StringVar(&r.html, "html", r.html, "HTML folder for HTTP/HTTPS connectors")
+	flagset.StringVar(&r.logLevel, "log-level", r.logLevel, "Sets the log level (debug, info, warn or error)")
+	flagset.BoolVar(&r.console, "console", r.console, "Runs as a console application rather than a service")
+	flagset.BoolVar(&r.debug, "debug", r.debug, "Enables detailed debugging logs")
 
 	return flagset
 }
@@ -79,7 +81,7 @@ func (cmd *Run) Description() string {
 }
 
 func (cmd *Run) Usage() string {
-	return "uhppoted-tunnel [--debug] [--console] [--lockfile <PID filepath>] --portal <UDP connection> --pipe <TCP connection>"
+	return "uhppoted-tunnel [--debug] [--console] [--lockfile <PID filepath>] --in <connection> --out <connection>"
 }
 
 func (cmd *Run) Help() {